requirements.txt
dash
dash-bootstrap-components
pandas
numpy
plotly
psycopg2
python-dotenv
databricks-sql-connector
databricks-sdk>=0.18.0
mlflow